Thinh

+++ bookingtrain/src/main/java/com/example/bookingtrain/repository/ObjectRepository.java
@@ -3,6 +3,5 @@ package com.example.bookingtrain.repository;
 import com.example.bookingtrain.model.Object;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ObjectRepository extends JpaRepository<Object, Long> {
-
+public interface ObjectRepository extends JpaRepository<Object, Integer> {
 }
